+++ app/Services/GuestCheckoutService.php
@@ -252,8 +252,8 @@ class GuestCheckoutService
     {
         $settings = Setting::allSettings();
 
-        $subtotal = collect($processedItems)->sum('total_price');
-        $discount = $couponCode ? $this->calculateGuestDiscount($subtotal, $couponCode) : 0;
+        $subtotal = (float) collect($processedItems)->sum('total_price');
+        $discount = $couponCode ? $this->calculateGuestDiscount($subtotal, $couponCode) : 0.0;
 
         $taxRate = (float) ($settings['tax_rate'] ?? 0);
         $tax = $subtotal * $taxRate;
@@ -263,7 +263,7 @@ class GuestCheckoutService
         $isFreeShipping = $subtotal >= $freeShippingThreshold;
 
         if ($isFreeShipping) {
-            $shipping = 0;
+            $shipping = 0.0;
         } else {
             $shipping = match ($shippingZone) {
                 'inside_dhaka' => (float) ($settings['shipping_cost_inside_dhaka'] ?? $settings['default_shipping_cost'] ?? 60),
@@ -272,7 +272,7 @@ class GuestCheckoutService
             };
         }
 
-        $total = max(0, $subtotal - $discount + $tax + $shipping);
+        $total = max(0.0, $subtotal - $discount + $tax + $shipping);
 
         return [
             'subtotal' => round($subtotal, 2),
